Retrofit implemented
- Api logic separated into specialized files
- Variables for using mock or real data added to AppModules
- Movie and User id changed from Int to Long to better management with
  the backend

[TODO]
*methods on Api files to fetch data from the backend
*methods on Impl files to give that data to components or viewmodels

+++ app/src/main/java/com/mooncowpines/kinostats/data/repositoryImpl/MockAuthRepositoryImpl.kt
@@ -34,7 +34,7 @@ class MockAuthRepositoryImpl : AuthRepository {
         )
     )
 
-    private var currentLoggedUserId: Int? = null
+    private var currentLoggedUserId: Long? = null
 
     override suspend fun login(email: String, pass: String): Boolean {
         delay(1500)
@@ -97,7 +97,7 @@ class MockAuthRepositoryImpl : AuthRepository {
         return false
     }
 
-    override suspend fun getUserById(userId: Int): User? {
+    override suspend fun getUserById(userId: Long): User? {
         delay(1500)
         return mockUsers.find { it.id == userId }
     }
