+++ app/src/main/java/com/mooncowpines/kinostats/domain/model/Review.kt
@@ -3,9 +3,9 @@ package com.mooncowpines.kinostats.domain.model
 import java.time.LocalDate
 
 data class Review(
-    val id: Int,
-    val movieId: Int,
-    val userId: Int?,
+    val id: Long,
+    val movieId: Long,
+    val userId: Long?,
     val rating: Float,
     val watchDate: LocalDate?,
     val reviewText: String
