+++ app/src/main/java/com/mooncowpines/kinostats/domain/repository/MovieRepository.kt
@@ -3,6 +3,6 @@ import com.mooncowpines.kinostats.domain.model.Movie
 
 interface MovieRepository {
     suspend fun getMovies(): List<Movie>
-    suspend fun getMovieById(id: Int): Movie?
+    suspend fun getMovieById(id: Long): Movie?
     suspend fun searchMovies(query: String): List<Movie>
 }
