+++ app/src/main/java/com/mooncowpines/kinostats/navigation/Routes.kt
@@ -9,14 +9,14 @@ sealed class Route(val path: String) {
     data object Profile  : Route("profile")
     data object Stats  : Route("stats")
     data object MovieDetail : Route("movie_detail/{movieId}") {
-        fun createRoute(movieId: Int) = "movie_detail/$movieId"
+        fun createRoute(movieId: Long) = "movie_detail/$movieId"
     }
     data object MovieLog : Route("movie_log/{movieId}") {
-        fun createRoute(movieId: Int) = "movie_log/$movieId"
+        fun createRoute(movieId: Long) = "movie_log/$movieId"
     }
 
     data object Review : Route("review/{movieId}") {
-        fun createRoute(movieId: Int) = "review/$movieId"
+        fun createRoute(movieId: Long) = "review/$movieId"
     }
 
     data object Search : Route("search/{query}") {
